cpu_input_boost: Allow configuration of each cluster's minimum frequency
Based on kdrag0n/proton_bluecross@64e4be6:

On some CPUs, it can be more efficient to use a higher base frequency
after all boosts have expired. For example, the low-power cluster on
sdm845 benefits from using a minimum frequency of 576000 KHz over 300000
KHz because it uses ~6% more power while offering nearly 2x the
performance. This is the minimum frequency used by Google, Qualcomm, and
OnePlus for good reason. It is actually possible for battery life to
improve when using this elevated base frequency because there are less
CPU frequency transitions (which are relatively expensive) and less
"ping-pong" effects where the CPU constantly switches between
frequencies in the 300000 KHz - 576000 KHz range. This has been observed
to occur during video playback in some cases.

This can also be used to bump the performance baseline for intensive
tasks such as gaming.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Rajan Palaniya <razorpaul@gmail.com>