qos: Replace cpumasks with atomic bitwise operations
This is expensive and unneeded, especially since the qos functions are
hot code paths. Mose of the cpumask functions use the bitmap API, which
is also more expensive than just doing some simple operations on a word.

Since we're operating with a CPU count that can fit within a word,
replace the expensive cpumask operations with plain bitwise operations
wherever possible to make the pm_qos framework more efficient.

Change-Id: Ida5f08d5c1e25aa790d814153f1c7fe7a5dd9850
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>