sched/tune: only force UX tasks to big cores
If we force top-app's boost value to be >0 we get excellent
jitter values due to top-app being rendered on big cores.

However along with the actual top app a load of crap ends up on
big cluster too. We can avoid this by checking if adj_score is 0,
which means that this app is actually the one that's currently on
the screen (or it's one of its child processes).

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
[alk3pInjection: Squash F1xy-kernels/SURYA_kernel@68ebacb]
Signed-off-by: Rajan Palaniya <razorpaul@gmail.com>