arm64: dts: 18821/19801: lower max brightness level to 2047
Brightness on the 1440p panels is actually capped at 2047. If one were
to set a value that's higher than that (e.g. 2048-4096), then the panel
will set a brightness level that's equivalent to `brightness % 2048`.

This leads to a very wacky user experience when booted into Android,
because if the user were to slide the brightness slider past the ~80%
mark, then the panel's brightness would wrap all the way back to 0,
which is not ideal.

This also leads to an even worse experience in recovery, because
recovery will set the brightness level to 2048. As a result, the panel
is set to zero backlight, and the user cannot see anything as a result.

Clearly, we are not supposed to be able to send a brightness level
higher than 2047, so let's lower the max brightness level accordingly.

Change-Id: I30a96aa0b25d2694be0c678afa136957fee2f95c